chore: relocate dummy classes
